feat(structure): Split codes according to their phases
